+++ drift/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Add support for FTS5 queries in the Dart API: `match`, `matchExp`, `highlight`, `snippet`, `bm25` and `rank` are now available as extension methods on FTS5 tables from `package:drift/extensions/fts5.dart` (#3843).
 - Breaking change (web): Remove `PortToChannel` extension on `dart:html` types. Use `WebPortToChannel` extension on `MessagePort` from `package:web` instead.
-- Web: Fix race to OPFS file handle in compatiblity check.
+- Web: Fix race to OPFS file handle in compatibility check.
 
 ## 2.34.4